+++ src/test/java/HolidayChecker.java
@@ -1,0 +1,7 @@
+import java.time.LocalDate;
+
+public class HolidayChecker {
+    public boolean isHoliday(final LocalDate targetDate) {
+        return true;
+    }
+}
