+++ src/test/java/OperationTimeChecker.java
@@ -6,8 +6,4 @@ public class OperationTimeChecker {
     public boolean isWeekend(final LocalDate targetDate) {
         return targetDate.getDayOfWeek() == DayOfWeek.SUNDAY || targetDate.getDayOfWeek() == DayOfWeek.SATURDAY;
     }
-
-    public boolean isHoliday(final LocalDate targetDate) {
-        return true;
-    }
 }
